Updating Android java build config Token storage solution Finishing Auth repository and UI layer

+++ data/network/public/src/commonMain/kotlin/app.campfire.network/models/ServerSettings.kt
@@ -14,7 +14,7 @@ data class ServerSettings(
   val storeMetadataWithItem: Boolean,
   val metadataFileFormat: String,
   val rateLimitLoginRequests: Int,
-  val rateLimitLoginWindow: Int,
+  val rateLimitLoginWindow: Long,
   val backupSchedule: String,
   val backupsToKeep: Int,
   val maxBackupSize: Int,
